pyproject.toml
[tool.poetry]
name = "povertool"
version = "0.1.0"
description = "Tools Collection"
authors = ["phoenix <hypofiasco@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
python-dotenv = "*"
Django = "3.1.7"
uvicorn = "^0.11.8"
gunicorn = "*"
gevent = "*"
celery = "*"
djangorestframework = "*"
coreapi = "*"
django-ratelimit = "^2.0"
django-cors-headers = "*"
django-environ = "*"
pyopenssl = "20.0.1"
pycryptodomex = "3.10.1"
httpx = "0.14"
passlib = "1.7.4"
pyDes = "2.0.1"
pyjwt = "2.0.1"
# aioredis = "1.3.1"
redis = "*"
sentry-sdk = "0.17.0"

[tool.poetry.dev-dependencies]
asynctest = "*"
bandit = "*"
black = "^19.10b0"
flake8 = "*"
isort = "5.1.4"
pytest = "*"
pytest-asyncio = "*"
pytest-httpx = "0.7.0"
taskipy = "*"

[[tool.poetry.source]]
url = "http://mirrors.aliyun.com/pypi/simple"
name = "aliyun"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
